requirements.txt
vkbottle
python-dotenv
black
sqlalchemy
psycopg2-binary
